PageSettingView.vue Data를 App.vue로 정리완료
